bug fixed: CoreData fatal error done
